Update Book.java
This class represents a book item that is sold in the retail store. The book should be identified by three instance variables: title (String), weight (double), price (double), and discount (double).


Implement the following methods for the Book class:

A constructor to initialize all instance variables except discount which is set to zero.

Getters all instance variables.

A setter method for discount. Note that discount will be entered as a percentage, you have to convert to its decimal number equivalent before used in a calculation. (e.g., 12% should converted to 0.12)

adjustPrice(): a method than takes an amount (of type double) and adds that amount of money to the book’s price. Note that the amount can be positive or negative as a book’s price may be increased or decreased.

priceAfterDiscount(): a method that calculates and returns the discounted book’s price after applying the discount. Note that this method should not change the value of the book’s price.

equals(): a method to test the equality of two books. Two books are considered to be equal if they have the same title, the same price and the same weight.

toString(): a method that returns a nicely formatted string description of the book.