Update Invoice.java
The Invoice class represents an invoice or a receipt. To keep the application simple, an invoice can contain at most three books. The Invoice class is identified by the following instance variables:

customer: a string that represents the customer's name.

books[]: an array type Book.


Implement the following methods for the Invoice class:

A constructor that takes as input only the customer's name and creates an invoice where the book[] is null.

getTotal(): a method that calculates and returns the total cost of that invoice (which is the sum of prices of all books in the invoice). Note that null books should not be included in the total cost calculations.

getDiscountedTotal(): a method that calculates and returns the total cost of that invoice (which is the sum of prices (after applying discount) of all the books in the invoice). Note that null books should not be included in the total cost calculations.

getTotalWeight(): a method that calculates and returns the total weight of all books in this invoice.

addBook(): a method that takes three input parameters to represent the book’s title, price and weight. The method then creates an object of type Book and assigns this object to one of the invoice's book[] array element. Note that the method should throw an exception if a book cannot be added if the invoice already has three books.

adjustPrice(): a method that takes two input, namely bookNum and priceAdjustment, where bookNum is either 1, 2 or 3. The method then adjusts the price of the given book. Note that the method should throw an exception if bookNum is invalid.

applyDiscount(): similar to adjustPrice(), this method sets discount of one of the invoice's books. The method takes two inputs, namely bookNum and discount, where bookNum is either 1, 2 or 3. The method then sets the specified book’s discount to the input discount value. The method should throw an exception if the input bookNum is invalid.

calculateTax(): a static method that is used to calculate the amount of sales tax to be paid. The method takes two input parameters both of type double. The first input parameter, called cost, represents the total cost to be taxed, while the second parameter, called taxRate, represents the tax rate in percentage. The method then calculates and returns the amount of tax to be paid using the following formula:


tax = (cost * taxRate)/100.


For example, if the cost is $60 and taxRate is 7.5 (i.e., 7.5%), the amount of tax is (60*7.5)/100 which is $4.5.


toString(): a method that returns a string representation of the invoice that includes the following:

customer's name,

details of the non-null invoice's books,

the total weight of all books in the invoice,

the invoice's total before and after discount,

the amount of tax to be paid where the tax is calculated based on the discounted invoice's total, and

the overall cost which is equal to discounted total + tax .